requirements.txt
python-telegram-bot==20.3
matplotlib==3.7.1
